refactor(geocoder): fix typescript errors
